Splitting Fault Injection tests to allow better parallelization.
