+++ e2e/test/FileUploadE2ETests.cs
@@ -50,50 +50,6 @@ namespace Microsoft.Azure.Devices.E2ETests
             await UploadFile(Client.TransportType.Http1, smallFile, true).ConfigureAwait(false);
         }
 
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task FileUploadSuccess_TcpLoss_Amqp()
-        {
-            string bigFile = await GetTestFileNameAsync(FileSizeBig).ConfigureAwait(false);
-
-            await UploadFileDisconnectTransport(Client.TransportType.Amqp_Tcp_Only,
-                bigFile,
-                FaultInjection.FaultType_Tcp,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec
-                ).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task FileUploadSuccess_Throttled_Amqp()
-        {
-            string smallFile = await GetTestFileNameAsync(FileSizeSmall).ConfigureAwait(false);
-
-            await UploadFileDisconnectTransport(Client.TransportType.Amqp_Tcp_Only,
-                smallFile,
-                FaultInjection.FaultType_Throttle,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec,
-                FaultInjection.DefaultDurationInSec
-                ).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task FileUploadSuccess_QuotaExceed_Amqp()
-        {
-            string smallFile = await GetTestFileNameAsync(FileSizeSmall).ConfigureAwait(false);
-
-            await UploadFileDisconnectTransport(Client.TransportType.Amqp_Tcp_Only,
-                smallFile,
-                FaultInjection.FaultType_QuotaExceeded,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec,
-                FaultInjection.DefaultDurationInSec
-                ).ConfigureAwait(false);
-        }
-
         private async Task UploadFile(Client.TransportType transport, string filename, bool x509auth = false)
         {
             TestDevice testDevice = await TestDevice.GetTestDeviceAsync(
@@ -160,57 +116,6 @@ namespace Microsoft.Azure.Devices.E2ETests
             }
             sw.Stop();
             return fileNotification;
-        }
-
-        private async Task UploadFileDisconnectTransport(
-            Client.TransportType transport, 
-            string filename, 
-            string faultType, 
-            string reason, 
-            int delayInSec, 
-            int durationInSec = 0, 
-            int retryDurationInMilliSec = FaultInjection.RecoveryTimeMilliseconds)
-        {
-            TestDevice testDevice = await TestDevice.GetTestDeviceAsync(DevicePrefix).ConfigureAwait(false);
-
-            using (DeviceClient deviceClient = DeviceClient.CreateFromConnectionString(testDevice.ConnectionString, transport))
-            using (ServiceClient serviceClient = ServiceClient.CreateFromConnectionString(Configuration.IoTHub.ConnectionString))
-            {
-                FileNotificationReceiver<FileNotification> notificationReceiver = serviceClient.GetFileNotificationReceiver();
-                deviceClient.OperationTimeoutInMilliseconds = (uint)retryDurationInMilliSec;
-
-                Task fileuploadTask;
-                Task<FileNotification> verifyTask;
-                using (FileStream fileStreamSource = new FileStream(filename, FileMode.Open, FileAccess.Read))
-                {
-                    verifyTask = VerifyFileNotification(notificationReceiver, testDevice.Id);
-                    fileuploadTask = deviceClient.UploadToBlobAsync(filename, fileStreamSource);
-
-                    try
-                    {
-                        await
-                            deviceClient.SendEventAsync(FaultInjection.ComposeErrorInjectionProperties(faultType, reason,
-                                delayInSec, durationInSec)).ConfigureAwait(false);
-                    }
-                    catch (Exception)
-                    {
-                        // catch and ignore exceptions resulted from error injection and continue to 
-                        // check result of the file upload status
-                    }
-
-                    await Task.WhenAll(fileuploadTask, verifyTask).ConfigureAwait(false);
-                }
-
-                FileNotification fileNotification = await verifyTask.ConfigureAwait(false);
-
-                Assert.IsNotNull(fileNotification, "FileNotification is not received.");
-                Assert.AreEqual(testDevice.Id + "/" + filename, fileNotification.BlobName, "Uploaded file name mismatch in notifications");
-                Assert.AreEqual(new FileInfo(filename).Length, fileNotification.BlobSizeInBytes, "Uploaded file size mismatch in notifications");
-                Assert.IsFalse(string.IsNullOrEmpty(fileNotification.BlobUri), "File notification blob uri is null or empty");
-
-                await deviceClient.CloseAsync().ConfigureAwait(false);
-                await serviceClient.CloseAsync().ConfigureAwait(false);
-            }
         }
 
         private async Task<string> GetTestFileNameAsync(int fileSize)
