+++ e2e/test/MessageReceiveE2ETests.cs
@@ -58,174 +58,6 @@ namespace Microsoft.Azure.Devices.E2ETests
         }
 
         [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_TcpConnectionLossReceiveRecovery_Amqp()
-        {
-            await ReceiveMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Amqp_Tcp_Only,
-                FaultInjection.FaultType_Tcp,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_TcpConnectionLossReceiveRecovery_AmqpWs()
-        {
-            await ReceiveMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Amqp_WebSocket_Only,
-                FaultInjection.FaultType_Tcp,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_TcpConnectionLossReceiveRecovery_Mqtt()
-        {
-            await ReceiveMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Mqtt_Tcp_Only,
-                FaultInjection.FaultType_Tcp,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec
-                ).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_TcpConnectionLossReceiveRecovery_MqttWs()
-        {
-            await ReceiveMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Mqtt_WebSocket_Only,
-                FaultInjection.FaultType_Tcp,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_AmqpConnectionLossReceiveRecovery_Amqp()
-        {
-            await ReceiveMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Amqp_Tcp_Only,
-                FaultInjection.FaultType_AmqpConn,
-                "",
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_AmqpConnectionLossReceiveRecovery_AmqpWs()
-        {
-            await ReceiveMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Amqp_WebSocket_Only,
-                FaultInjection.FaultType_AmqpConn, "",
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_AmqpSessionLossReceiveRecovery_Amqp()
-        {
-            await ReceiveMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Amqp_Tcp_Only,
-                FaultInjection.FaultType_AmqpSess,
-                "",
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_AmqpSessionLossReceiveRecovery_AmqpWs()
-        {
-            await ReceiveMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Amqp_WebSocket_Only,
-                FaultInjection.FaultType_AmqpSess,
-                "",
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_AmqpC2DLinkDropReceiveRecovery_Amqp()
-        {
-            await ReceiveMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Amqp_Tcp_Only,
-                FaultInjection.FaultType_AmqpC2D,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_AmqpC2DLinkDropReceiveRecovery_AmqpWs()
-        {
-            await ReceiveMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Amqp_WebSocket_Only,
-                FaultInjection.FaultType_AmqpD2C,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_GracefulShutdownReceiveRecovery_Amqp()
-        {
-            await ReceiveMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Amqp_Tcp_Only,
-                FaultInjection.FaultType_GracefulShutdownAmqp,
-                FaultInjection.FaultCloseReason_Bye,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_GracefulShutdownReceiveRecovery_AmqpWs()
-        {
-            await ReceiveMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Amqp_WebSocket_Only,
-                FaultInjection.FaultType_GracefulShutdownAmqp,
-                FaultInjection.FaultCloseReason_Bye,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_GracefulShutdownReceiveRecovery_Mqtt()
-        {
-            await ReceiveMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Mqtt_Tcp_Only,
-                FaultInjection.FaultType_GracefulShutdownMqtt,
-                FaultInjection.FaultCloseReason_Bye,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_GracefulShutdownReceiveRecovery_MqttWs()
-        {
-            await ReceiveMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Mqtt_WebSocket_Only,
-                FaultInjection.FaultType_GracefulShutdownMqtt,
-                FaultInjection.FaultCloseReason_Bye,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
         [ExpectedException(typeof(TimeoutException))]
         public async Task Message_TimeOutReachedResponse()
         {
@@ -375,55 +207,6 @@ namespace Microsoft.Azure.Devices.E2ETests
                 await VerifyReceivedC2DMessage(transport, deviceClient, payload, p1Value).ConfigureAwait(false);
                 await deviceClient.CloseAsync().ConfigureAwait(false);
                 await serviceClient.CloseAsync().ConfigureAwait(false);
-            }
-        }
-
-        private async Task ReceiveMessageRecovery(
-            TestDeviceType type,
-            Client.TransportType transport,
-            string faultType,
-            string reason,
-            int delayInSec)
-        {
-            using (ServiceClient serviceClient = ServiceClient.CreateFromConnectionString(Configuration.IoTHub.ConnectionString))
-            {
-                Func<DeviceClient, TestDevice, Task> init = async (deviceClient, testDevice) =>
-                {
-                    await serviceClient.OpenAsync().ConfigureAwait(false);
-
-                    if (transport == Client.TransportType.Mqtt_Tcp_Only ||
-                        transport == Client.TransportType.Mqtt_WebSocket_Only)
-                    {
-                        // Dummy ReceiveAsync to ensure mqtt subscription registration before SendAsync() is called on service client.
-                        await deviceClient.ReceiveAsync(TimeSpan.FromSeconds(2)).ConfigureAwait(false);
-                    }
-                };
-
-                Func<DeviceClient, TestDevice, Task> testOperation = async (deviceClient, testDevice) =>
-                {
-                    string payload, messageId, p1Value;
-                    await serviceClient.SendAsync(
-                        testDevice.Id,
-                        ComposeC2DTestMessage(out payload, out messageId, out p1Value)).ConfigureAwait(false);
-                    await VerifyReceivedC2DMessage(transport, deviceClient, payload, p1Value).ConfigureAwait(false);
-                };
-
-                Func<Task> cleanupOperation = () =>
-                {
-                    return serviceClient.CloseAsync();
-                };
-
-                await FaultInjection.TestErrorInjectionTemplate(
-                    DevicePrefix,
-                    type,
-                    transport,
-                    faultType,
-                    reason,
-                    delayInSec,
-                    FaultInjection.DefaultDurationInSec,
-                    init,
-                    testOperation,
-                    cleanupOperation).ConfigureAwait(false);
             }
         }
