+++ e2e/test/MessageReceiveFaultInjectionTests.cs
@@ -1,0 +1,315 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using Microsoft.Azure.Devices.Client;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using System;
+using System.Diagnostics;
+using System.Diagnostics.Tracing;
+using System.Linq;
+using System.Net;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Microsoft.Azure.Devices.E2ETests
+{
+    [TestClass]
+    [TestCategory("IoTHub-E2E")]
+    [TestCategory("IoTHub-FaultInjection")]
+    public partial class MessageReceiveFaultInjectionTests : IDisposable
+    {
+        private readonly string DevicePrefix = $"E2E_{nameof(MessageReceiveFaultInjectionTests)}_";
+        private static TestLogging _log = TestLogging.GetInstance();
+
+        private readonly ConsoleEventListener _listener;
+
+        public MessageReceiveFaultInjectionTests()
+        {
+            _listener = TestConfig.StartEventListener();
+        }
+
+        [TestMethod]
+        public async Task Message_TcpConnectionLossReceiveRecovery_Amqp()
+        {
+            await ReceiveMessageRecovery(
+                TestDeviceType.Sasl,
+                Client.TransportType.Amqp_Tcp_Only,
+                FaultInjection.FaultType_Tcp,
+                FaultInjection.FaultCloseReason_Boom,
+                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
+        }
+
+        [TestMethod]
+        public async Task Message_TcpConnectionLossReceiveRecovery_AmqpWs()
+        {
+            await ReceiveMessageRecovery(
+                TestDeviceType.Sasl,
+                Client.TransportType.Amqp_WebSocket_Only,
+                FaultInjection.FaultType_Tcp,
+                FaultInjection.FaultCloseReason_Boom,
+                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
+        }
+
+        [TestMethod]
+        public async Task Message_TcpConnectionLossReceiveRecovery_Mqtt()
+        {
+            await ReceiveMessageRecovery(
+                TestDeviceType.Sasl,
+                Client.TransportType.Mqtt_Tcp_Only,
+                FaultInjection.FaultType_Tcp,
+                FaultInjection.FaultCloseReason_Boom,
+                FaultInjection.DefaultDelayInSec
+                ).ConfigureAwait(false);
+        }
+
+        [TestMethod]
+        public async Task Message_TcpConnectionLossReceiveRecovery_MqttWs()
+        {
+            await ReceiveMessageRecovery(
+                TestDeviceType.Sasl,
+                Client.TransportType.Mqtt_WebSocket_Only,
+                FaultInjection.FaultType_Tcp,
+                FaultInjection.FaultCloseReason_Boom,
+                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
+        }
+
+        [TestMethod]
+        public async Task Message_AmqpConnectionLossReceiveRecovery_Amqp()
+        {
+            await ReceiveMessageRecovery(
+                TestDeviceType.Sasl,
+                Client.TransportType.Amqp_Tcp_Only,
+                FaultInjection.FaultType_AmqpConn,
+                "",
+                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
+        }
+
+        [TestMethod]
+        public async Task Message_AmqpConnectionLossReceiveRecovery_AmqpWs()
+        {
+            await ReceiveMessageRecovery(
+                TestDeviceType.Sasl,
+                Client.TransportType.Amqp_WebSocket_Only,
+                FaultInjection.FaultType_AmqpConn, "",
+                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
+        }
+
+        [TestMethod]
+        public async Task Message_AmqpSessionLossReceiveRecovery_Amqp()
+        {
+            await ReceiveMessageRecovery(
+                TestDeviceType.Sasl,
+                Client.TransportType.Amqp_Tcp_Only,
+                FaultInjection.FaultType_AmqpSess,
+                "",
+                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
+        }
+
+        [TestMethod]
+        public async Task Message_AmqpSessionLossReceiveRecovery_AmqpWs()
+        {
+            await ReceiveMessageRecovery(
+                TestDeviceType.Sasl,
+                Client.TransportType.Amqp_WebSocket_Only,
+                FaultInjection.FaultType_AmqpSess,
+                "",
+                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
+        }
+
+        [TestMethod]
+        public async Task Message_AmqpC2DLinkDropReceiveRecovery_Amqp()
+        {
+            await ReceiveMessageRecovery(
+                TestDeviceType.Sasl,
+                Client.TransportType.Amqp_Tcp_Only,
+                FaultInjection.FaultType_AmqpC2D,
+                FaultInjection.FaultCloseReason_Boom,
+                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
+        }
+
+        [TestMethod]
+        public async Task Message_AmqpC2DLinkDropReceiveRecovery_AmqpWs()
+        {
+            await ReceiveMessageRecovery(
+                TestDeviceType.Sasl,
+                Client.TransportType.Amqp_WebSocket_Only,
+                FaultInjection.FaultType_AmqpD2C,
+                FaultInjection.FaultCloseReason_Boom,
+                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
+        }
+
+        [TestMethod]
+        public async Task Message_GracefulShutdownReceiveRecovery_Amqp()
+        {
+            await ReceiveMessageRecovery(
+                TestDeviceType.Sasl,
+                Client.TransportType.Amqp_Tcp_Only,
+                FaultInjection.FaultType_GracefulShutdownAmqp,
+                FaultInjection.FaultCloseReason_Bye,
+                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
+        }
+
+        [TestMethod]
+        public async Task Message_GracefulShutdownReceiveRecovery_AmqpWs()
+        {
+            await ReceiveMessageRecovery(
+                TestDeviceType.Sasl,
+                Client.TransportType.Amqp_WebSocket_Only,
+                FaultInjection.FaultType_GracefulShutdownAmqp,
+                FaultInjection.FaultCloseReason_Bye,
+                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
+        }
+
+        [TestMethod]
+        public async Task Message_GracefulShutdownReceiveRecovery_Mqtt()
+        {
+            await ReceiveMessageRecovery(
+                TestDeviceType.Sasl,
+                Client.TransportType.Mqtt_Tcp_Only,
+                FaultInjection.FaultType_GracefulShutdownMqtt,
+                FaultInjection.FaultCloseReason_Bye,
+                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
+        }
+
+        [TestMethod]
+        public async Task Message_GracefulShutdownReceiveRecovery_MqttWs()
+        {
+            await ReceiveMessageRecovery(
+                TestDeviceType.Sasl,
+                Client.TransportType.Mqtt_WebSocket_Only,
+                FaultInjection.FaultType_GracefulShutdownMqtt,
+                FaultInjection.FaultCloseReason_Bye,
+                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
+        }
+
+        private Client.Message ComposeD2CTestMessage(out string payload, out string p1Value)
+        {
+            payload = Guid.NewGuid().ToString();
+            p1Value = Guid.NewGuid().ToString();
+
+            _log.WriteLine($"{nameof(ComposeD2CTestMessage)}: payload='{payload}' p1Value='{p1Value}'");
+
+            return new Client.Message(Encoding.UTF8.GetBytes(payload))
+            {
+                Properties = { ["property1"] = p1Value }
+            };
+        }
+
+        private Message ComposeC2DTestMessage(out string payload, out string messageId, out string p1Value)
+        {
+            payload = Guid.NewGuid().ToString();
+            messageId = Guid.NewGuid().ToString();
+            p1Value = Guid.NewGuid().ToString();
+
+            _log.WriteLine($"{nameof(ComposeC2DTestMessage)}: payload='{payload}' messageId='{messageId}' p1Value='{p1Value}'");
+
+            return new Message(Encoding.UTF8.GetBytes(payload))
+            {
+                MessageId = messageId,
+                Properties = { ["property1"] = p1Value }
+            };
+        }
+
+        private async Task VerifyReceivedC2DMessage(Client.TransportType transport, DeviceClient dc, string payload, string p1Value)
+        {
+            var wait = true;
+
+            Stopwatch sw = new Stopwatch();
+            sw.Start();
+            while (wait)
+            {
+                Client.Message receivedMessage;
+
+                if (transport == Client.TransportType.Http1)
+                {
+                    // Long-polling is not supported in http
+                    receivedMessage = await dc.ReceiveAsync().ConfigureAwait(false);
+                }
+                else
+                {
+                    receivedMessage = await dc.ReceiveAsync(TimeSpan.FromSeconds(1)).ConfigureAwait(false);
+                }
+
+                if (receivedMessage != null)
+                {
+                    string messageData = Encoding.ASCII.GetString(receivedMessage.GetBytes());
+                    Assert.AreEqual(messageData, payload);
+
+                    Assert.AreEqual(receivedMessage.Properties.Count, 1);
+                    var prop = receivedMessage.Properties.Single();
+                    Assert.AreEqual(prop.Key, "property1");
+                    Assert.AreEqual(prop.Value, p1Value);
+
+                    await dc.CompleteAsync(receivedMessage).ConfigureAwait(false);
+                    wait = false;
+                }
+
+                if (sw.Elapsed.TotalSeconds > 5)
+                {
+                    throw new TimeoutException("Test is running longer than expected.");
+                }
+            }
+
+            sw.Stop();
+        }
+
+        private async Task ReceiveMessageRecovery(
+            TestDeviceType type,
+            Client.TransportType transport,
+            string faultType,
+            string reason,
+            int delayInSec)
+        {
+            using (ServiceClient serviceClient = ServiceClient.CreateFromConnectionString(Configuration.IoTHub.ConnectionString))
+            {
+                Func<DeviceClient, TestDevice, Task> init = async (deviceClient, testDevice) =>
+                {
+                    await serviceClient.OpenAsync().ConfigureAwait(false);
+
+                    if (transport == Client.TransportType.Mqtt_Tcp_Only ||
+                        transport == Client.TransportType.Mqtt_WebSocket_Only)
+                    {
+                        // Dummy ReceiveAsync to ensure mqtt subscription registration before SendAsync() is called on service client.
+                        await deviceClient.ReceiveAsync(TimeSpan.FromSeconds(2)).ConfigureAwait(false);
+                    }
+                };
+
+                Func<DeviceClient, TestDevice, Task> testOperation = async (deviceClient, testDevice) =>
+                {
+                    string payload, messageId, p1Value;
+                    await serviceClient.SendAsync(
+                        testDevice.Id,
+                        ComposeC2DTestMessage(out payload, out messageId, out p1Value)).ConfigureAwait(false);
+                    await VerifyReceivedC2DMessage(transport, deviceClient, payload, p1Value).ConfigureAwait(false);
+                };
+
+                Func<Task> cleanupOperation = () =>
+                {
+                    return serviceClient.CloseAsync();
+                };
+
+                await FaultInjection.TestErrorInjectionTemplate(
+                    DevicePrefix,
+                    type,
+                    transport,
+                    faultType,
+                    reason,
+                    delayInSec,
+                    FaultInjection.DefaultDurationInSec,
+                    init,
+                    testOperation,
+                    cleanupOperation).ConfigureAwait(false);
+            }
+        }
+
+        public void Dispose()
+        {
+            Dispose(true);
+            GC.SuppressFinalize(this);
+        }
+
+        protected virtual void Dispose(bool disposing)
+        {
+        }
+    }
+}
