+++ e2e/test/MessageSendE2ETests.cs
@@ -131,330 +131,6 @@ namespace Microsoft.Azure.Devices.E2ETests
         }
 
         [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_TcpConnectionLossSendRecovery_Amqp()
-        {
-            // NOTE: FaultInjection will not actually terminate the TCP layer.
-            await SendMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Amqp_Tcp_Only,
-                FaultInjection.FaultType_Tcp,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_TcpConnectionLossSendRecovery_AmqpWs()
-        {
-            // NOTE: FaultInjection will not actually terminate the TCP layer.
-            await SendMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Amqp_WebSocket_Only,
-                FaultInjection.FaultType_Tcp,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_TcpConnectionLossSendRecovery_Mqtt()
-        {
-            await SendMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Mqtt_Tcp_Only,
-                FaultInjection.FaultType_Tcp,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_TcpConnectionLossSendRecovery_MqttWs()
-        {
-            await SendMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Mqtt_WebSocket_Only,
-                FaultInjection.FaultType_Tcp,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_AmqpConnectionLossSendRecovery_Amqp()
-        {
-            await SendMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Amqp_Tcp_Only,
-                FaultInjection.FaultType_AmqpConn,
-                "",
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_AmqpConnectionLossSendRecovery_AmqpWs()
-        {
-            await SendMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Amqp_WebSocket_Only,
-                FaultInjection.FaultType_AmqpConn,
-                "",
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_AmqpSessionLossSendRecovery_Amqp()
-        {
-            await SendMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Amqp_Tcp_Only,
-                FaultInjection.FaultType_AmqpSess,
-                "",
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_AmqpSessionLossSendRecovery_AmqpWs()
-        {
-            await SendMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Amqp_WebSocket_Only,
-                FaultInjection.FaultType_AmqpSess,
-                "",
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_AmqpD2CLinkDropSendRecovery_Amqp()
-        {
-            await SendMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Amqp_Tcp_Only,
-                FaultInjection.FaultType_AmqpD2C,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_AmqpD2CLinkDropSendRecovery_AmqpWs()
-        {
-            await SendMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Amqp_WebSocket_Only,
-                FaultInjection.FaultType_AmqpD2C,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_ThrottledConnectionRecovery_Amqp()
-        {
-            await SendMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Amqp_Tcp_Only,
-                FaultInjection.FaultType_Throttle,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec,
-                FaultInjection.DefaultDurationInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_ThrottledConnectionRecovery_AmqpWs()
-        {
-            await SendMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Amqp_WebSocket_Only,
-                FaultInjection.FaultType_Throttle,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec,
-                FaultInjection.DefaultDurationInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        [ExpectedException(typeof(TimeoutException))]
-        public async Task Message_ThrottledConnectionLongTimeNoRecovery_Amqp()
-        {
-            await SendMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Amqp_Tcp_Only,
-                FaultInjection.FaultType_Throttle,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec,
-                FaultInjection.DefaultDurationInSec,
-                FaultInjection.ShortRetryInMilliSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        [ExpectedException(typeof(TimeoutException))]
-        public async Task Message_ThrottledConnectionLongTimeNoRecovery_AmqpWs()
-        {
-            await SendMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Amqp_WebSocket_Only,
-                FaultInjection.FaultType_Throttle,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec,
-                FaultInjection.DefaultDurationInSec,
-                FaultInjection.ShortRetryInMilliSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        [ExpectedException(typeof(QuotaExceededException))]
-        public async Task Message_ThrottledConnectionLongTimeNoRecovery_Http()
-        {
-            await SendMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Http1,
-                FaultInjection.FaultType_QuotaExceeded,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec,
-                FaultInjection.DefaultDurationInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        [ExpectedException(typeof(DeviceMaximumQueueDepthExceededException))]
-        public async Task Message_QuotaExceededRecovery_Amqp()
-        {
-            await SendMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Amqp_Tcp_Only,
-                FaultInjection.FaultType_QuotaExceeded,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec,
-                FaultInjection.DefaultDurationInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        [ExpectedException(typeof(DeviceMaximumQueueDepthExceededException))]
-        public async Task Message_QuotaExceededRecovery_AmqpWs()
-        {
-            await SendMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Amqp_WebSocket_Only,
-                FaultInjection.FaultType_QuotaExceeded,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec,
-                FaultInjection.DefaultDurationInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        [ExpectedException(typeof(QuotaExceededException))]
-        public async Task Message_QuotaExceededRecovery_Http()
-        {
-            await SendMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Http1,
-                FaultInjection.FaultType_QuotaExceeded,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec,
-                FaultInjection.DefaultDurationInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        [ExpectedException(typeof(UnauthorizedException))]
-        public async Task Message_AuthenticationRecovery_Amqp()
-        {
-            await SendMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Amqp_Tcp_Only,
-                FaultInjection.FaultType_Auth,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec,
-                FaultInjection.DefaultDurationInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        [ExpectedException(typeof(UnauthorizedException))]
-        public async Task Message_AuthenticationRecovery_AmqpWs()
-        {
-            await SendMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Amqp_WebSocket_Only,
-                FaultInjection.FaultType_Auth,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec,
-                FaultInjection.DefaultDurationInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        [ExpectedException(typeof(UnauthorizedException))]
-        public async Task Message_AuthenticationRecovery_Http()
-        {
-            await SendMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Http1,
-                FaultInjection.FaultType_Auth,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec,
-                FaultInjection.DefaultDurationInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_GracefulShutdownSendRecovery_Amqp()
-        {
-            await SendMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Amqp_Tcp_Only,
-                FaultInjection.FaultType_GracefulShutdownAmqp,
-                FaultInjection.FaultCloseReason_Bye,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_GracefulShutdownSendRecovery_AmqpWs()
-        {
-            await SendMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Amqp_WebSocket_Only,
-                FaultInjection.FaultType_GracefulShutdownAmqp,
-                FaultInjection.FaultCloseReason_Bye,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_GracefulShutdownSendRecovery_Mqtt()
-        {
-            await SendMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Mqtt_Tcp_Only,
-                FaultInjection.FaultType_GracefulShutdownMqtt,
-                FaultInjection.FaultCloseReason_Bye,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Message_GracefulShutdownSendRecovery_MqttWs()
-        {
-            await SendMessageRecovery(
-                TestDeviceType.Sasl,
-                Client.TransportType.Mqtt_WebSocket_Only,
-                FaultInjection.FaultType_GracefulShutdownMqtt,
-                FaultInjection.FaultCloseReason_Bye,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
         [ExpectedException(typeof(TimeoutException))]
         public async Task Message_TimeOutReachedResponse()
         {
@@ -547,49 +223,6 @@ namespace Microsoft.Azure.Devices.E2ETests
             };
         }
 
-        private async Task VerifyReceivedC2DMessage(Client.TransportType transport, DeviceClient dc, string payload, string p1Value)
-        {
-            var wait = true;
-
-            Stopwatch sw = new Stopwatch();
-            sw.Start();
-            while (wait)
-            {
-                Client.Message receivedMessage;
-
-                if (transport == Client.TransportType.Http1)
-                {
-                    // Long-polling is not supported in http
-                    receivedMessage = await dc.ReceiveAsync().ConfigureAwait(false);
-                }
-                else
-                {
-                    receivedMessage = await dc.ReceiveAsync(TimeSpan.FromSeconds(1)).ConfigureAwait(false);
-                }
-
-                if (receivedMessage != null)
-                {
-                    string messageData = Encoding.ASCII.GetString(receivedMessage.GetBytes());
-                    Assert.AreEqual(messageData, payload);
-
-                    Assert.AreEqual(receivedMessage.Properties.Count, 1);
-                    var prop = receivedMessage.Properties.Single();
-                    Assert.AreEqual(prop.Key, "property1");
-                    Assert.AreEqual(prop.Value, p1Value);
-
-                    await dc.CompleteAsync(receivedMessage).ConfigureAwait(false);
-                    wait = false;
-                }
-
-                if (sw.Elapsed.TotalSeconds > 5)
-                {
-                    throw new TimeoutException("Test is running longer than expected.");
-                }
-            }
-
-            sw.Stop();
-        }
-
         private async Task SendSingleMessage(TestDeviceType type, Client.TransportType transport)
         {
             TestDevice testDevice = await TestDevice.GetTestDeviceAsync(DevicePrefix, type).ConfigureAwait(false);
@@ -644,53 +277,6 @@ namespace Microsoft.Azure.Devices.E2ETests
                 await moduleClient.SendEventAsync(testMessage).ConfigureAwait(false);
                 await moduleClient.CloseAsync().ConfigureAwait(false);
             }
-        }
-
-        internal async Task SendMessageRecovery(
-            TestDeviceType type,
-            Client.TransportType transport,
-            string faultType, 
-            string reason, 
-            int delayInSec, 
-            int durationInSec = FaultInjection.DefaultDurationInSec, 
-            int retryDurationInMilliSec = FaultInjection.RecoveryTimeMilliseconds)
-        {
-            EventHubTestListener testListener = null;
-
-            Func<DeviceClient, TestDevice, Task> init = async (deviceClient, testDevice) =>
-            {
-                testListener = await EventHubTestListener.CreateListener(testDevice.Id).ConfigureAwait(false);
-                deviceClient.OperationTimeoutInMilliseconds = (uint)retryDurationInMilliSec;
-            };
-
-            Func<DeviceClient, TestDevice, Task> testOperation = async (deviceClient, testDevice) =>
-            {
-                string payload, p1Value;
-
-                Client.Message testMessage = ComposeD2CTestMessage(out payload, out p1Value);
-                await deviceClient.SendEventAsync(testMessage).ConfigureAwait(false);
-
-                bool isReceived = false;
-                isReceived = await testListener.WaitForMessage(testDevice.Id, payload, p1Value).ConfigureAwait(false);
-                Assert.IsTrue(isReceived);
-            };
-
-            Func<Task> cleanupOperation = () =>
-            {
-                return testListener.CloseAsync();
-            };
-            
-            await FaultInjection.TestErrorInjectionTemplate(
-                DevicePrefix,
-                type,
-                transport,
-                faultType,
-                reason,
-                delayInSec,
-                durationInSec,
-                init,
-                testOperation,
-                cleanupOperation).ConfigureAwait(false);
         }
         
         public void Dispose()
