+++ e2e/test/MethodE2ETests.cs
@@ -101,181 +101,6 @@ namespace Microsoft.Azure.Devices.E2ETests
             await SendMethodAndRespond(Client.TransportType.Amqp_WebSocket_Only, SetDeviceReceiveMethodDefaultHandler).ConfigureAwait(false);
         }
 
-        [Ignore] // TODO: #558
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Method_DeviceReceivesMethodAndResponseRecovery_Mqtt()
-        {
-            await SendMethodAndRespondRecovery(Client.TransportType.Mqtt_Tcp_Only, 
-                FaultInjection.FaultType_Tcp,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [Ignore] // TODO: #558
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Method_DeviceReceivesMethodAndResponseRecovery_MqttWs()
-        {
-            await SendMethodAndRespondRecovery(Client.TransportType.Mqtt_WebSocket_Only, 
-                FaultInjection.FaultType_Tcp,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [Ignore] // TODO: #558
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Method_DeviceMethodGracefulShutdownRecovery_Mqtt()
-        {
-            await SendMethodAndRespondRecovery(Client.TransportType.Mqtt_Tcp_Only,
-                FaultInjection.FaultType_GracefulShutdownMqtt,
-                FaultInjection.FaultCloseReason_Bye,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [Ignore] // TODO: #558
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Method_DeviceMethodGracefulShutdownRecovery_MqttWs()
-        {
-            await SendMethodAndRespondRecovery(Client.TransportType.Mqtt_WebSocket_Only,
-                FaultInjection.FaultType_GracefulShutdownMqtt,
-                FaultInjection.FaultCloseReason_Bye,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [Ignore] //TODO: #571
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Method_DeviceMethodTcpConnRecovery_Amqp()
-        {
-            await SendMethodAndRespondRecovery(Client.TransportType.Amqp_Tcp_Only,
-                FaultInjection.FaultType_Tcp,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Method_DeviceMethodTcpConnRecovery_AmqpWs()
-        {
-            await SendMethodAndRespondRecovery(Client.TransportType.Amqp_WebSocket_Only,
-                FaultInjection.FaultType_Tcp,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [Ignore] //TODO: #571
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Method_DeviceMethodAmqpConnLostRecovery_Amqp()
-        {
-            await SendMethodAndRespondRecovery(Client.TransportType.Amqp_Tcp_Only,
-                FaultInjection.FaultType_AmqpConn,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [Ignore] //TODO: #571
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Method_DeviceMethodAmqpConnLostRecovery_AmqpWs()
-        {
-            await SendMethodAndRespondRecovery(Client.TransportType.Amqp_WebSocket_Only, 
-                FaultInjection.FaultType_AmqpConn,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [Ignore] //TODO: #571
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Method_DeviceMethodSessionLostRecovery_Amqp()
-        {
-            await SendMethodAndRespondRecovery(Client.TransportType.Amqp_Tcp_Only,
-                FaultInjection.FaultType_AmqpSess,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [Ignore] //TODO: #571
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Method_DeviceMethodSessionLostRecovery_AmqpWs()
-        {
-            await SendMethodAndRespondRecovery(Client.TransportType.Amqp_WebSocket_Only,
-                FaultInjection.FaultType_AmqpSess,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [Ignore] //TODO: #194 Test intermittently failing on Windows.
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Method_DeviceMethodReqLinkDropRecovery_Amqp()
-        {
-            await SendMethodAndRespondRecovery(Client.TransportType.Amqp_Tcp_Only,
-                FaultInjection.FaultType_AmqpMethodReq,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [Ignore] //TODO: #194 Test intermittently failing on Windows.
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Method_DeviceMethodReqLinkDropRecovery_AmqpWs()
-        {
-            await SendMethodAndRespondRecovery(Client.TransportType.Amqp_WebSocket_Only,
-                FaultInjection.FaultType_AmqpMethodReq,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [Ignore] //TODO: #194 Test intermittently failing on Windows.
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Method_DeviceMethodRespLinkDropRecovery_Amqp()
-        {
-            await SendMethodAndRespondRecovery(Client.TransportType.Amqp_Tcp_Only,
-                FaultInjection.FaultType_AmqpMethodResp,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [Ignore] //TODO: #194 Test intermittently failing on Windows.
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Method_DeviceMethodRespLinkDropRecovery_AmqpWs()
-        {
-            await SendMethodAndRespondRecovery(Client.TransportType.Amqp_WebSocket_Only,
-                FaultInjection.FaultType_AmqpMethodResp,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [Ignore] //TODO: #571
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Method_DeviceMethodGracefulShutdownRecovery_Amqp()
-        {
-            await SendMethodAndRespondRecovery(Client.TransportType.Amqp_Tcp_Only,
-                FaultInjection.FaultType_GracefulShutdownAmqp,
-                FaultInjection.FaultCloseReason_Bye,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [Ignore] //TODO: #571
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Method_DeviceMethodGracefulShutdownRecovery_AmqpWs()
-        {
-            await SendMethodAndRespondRecovery(Client.TransportType.Amqp_WebSocket_Only,
-                FaultInjection.FaultType_GracefulShutdownAmqp,
-                FaultInjection.FaultCloseReason_Bye,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
         private async Task ServiceSendMethodAndVerifyResponse(string deviceName, string methodName, string respJson, string reqJson)
         {
             using (ServiceClient serviceClient = ServiceClient.CreateFromConnectionString(Configuration.IoTHub.ConnectionString))
@@ -369,30 +194,6 @@ namespace Microsoft.Azure.Devices.E2ETests
 
                 await deviceClient.CloseAsync().ConfigureAwait(false);
             }
-        }
-
-        private async Task SendMethodAndRespondRecovery(Client.TransportType transport, string faultType, string reason, int delayInSec)
-        {
-            Func<DeviceClient, TestDevice, Task> testOperation = async (deviceClient, testDevice) =>
-            {
-                Task methodReceivedTask = await SetDeviceReceiveMethod(deviceClient).ConfigureAwait(false);
-
-                await Task.WhenAll(
-                    ServiceSendMethodAndVerifyResponse(testDevice.Id, MethodName, DeviceResponseJson, ServiceRequestJson),
-                    methodReceivedTask).ConfigureAwait(false);
-            };
-
-            await FaultInjection.TestErrorInjectionTemplate(
-                DevicePrefix,
-                TestDeviceType.Sasl,
-                transport,
-                faultType,
-                reason,
-                delayInSec,
-                FaultInjection.DefaultDelayInSec,
-                (d, t) => { return Task.FromResult<bool>(false); },
-                testOperation,
-                () => { return Task.FromResult<bool>(false); }).ConfigureAwait(false);
         }
 
         public void Dispose()
