+++ e2e/test/TwinE2ETests.cs
@@ -48,98 +48,6 @@ namespace Microsoft.Azure.Devices.E2ETests
             await Twin_DeviceSetsReportedPropertyAndGetsItBack(Client.TransportType.Amqp_WebSocket_Only).ConfigureAwait(false);
         }
 
-        [Ignore] // TODO: #558
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Twin_DeviceReportedPropertiesTcpConnRecovery_Mqtt()
-        {
-            await Twin_DeviceReportedPropertiesRecovery(Client.TransportType.Mqtt_Tcp_Only, 
-                FaultInjection.FaultType_Tcp,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [Ignore] // TODO: #558
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Twin_DeviceReportedPropertiesTcpConnRecovery_MqttWs()
-        {
-            await Twin_DeviceReportedPropertiesRecovery(Client.TransportType.Mqtt_WebSocket_Only, 
-                FaultInjection.FaultType_Tcp,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Twin_DeviceReportedPropertiesTcpConnRecovery_Amqp()
-        {
-            await Twin_DeviceReportedPropertiesRecovery(Client.TransportType.Amqp_Tcp_Only, 
-                FaultInjection.FaultType_Tcp,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-#if NETCOREAPP2_0
-        // TODO: #302 In NetCoreApp2.0 the test is failing with TimeoutException.
-        [Ignore]
-#endif
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Twin_DeviceReportedPropertiesTcpConnRecovery_AmqpWs()
-        {
-            await Twin_DeviceReportedPropertiesRecovery(Client.TransportType.Amqp_WebSocket_Only, 
-                FaultInjection.FaultType_Tcp,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [Ignore] // TODO: #558
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Twin_DeviceReportedPropertiesGracefulShutdownRecovery_Mqtt()
-        {
-            await Twin_DeviceReportedPropertiesRecovery(Client.TransportType.Mqtt_Tcp_Only,
-                FaultInjection.FaultType_GracefulShutdownMqtt,
-                FaultInjection.FaultCloseReason_Bye,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [Ignore] // TODO: #558
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Twin_DeviceReportedPropertiesGracefulShutdownRecovery_MqttWs()
-        {
-            await Twin_DeviceReportedPropertiesRecovery(Client.TransportType.Mqtt_WebSocket_Only,
-                FaultInjection.FaultType_GracefulShutdownMqtt,
-                FaultInjection.FaultCloseReason_Bye,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Twin_DeviceReportedPropertiesGracefulShutdownRecovery_Amqp()
-        {
-            await Twin_DeviceReportedPropertiesRecovery(Client.TransportType.Amqp_Tcp_Only,
-                FaultInjection.FaultType_GracefulShutdownAmqp,
-                FaultInjection.FaultCloseReason_Bye,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-#if NETCOREAPP2_0
-        // TODO: #302 In NetCoreApp2.0 the test is failing with TimeoutException.
-        [Ignore]
-#endif
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Twin_DeviceReportedPropertiesGracefulShutdownRecovery_AmqpWs()
-        {
-            await Twin_DeviceReportedPropertiesRecovery(Client.TransportType.Amqp_WebSocket_Only,
-                FaultInjection.FaultType_GracefulShutdownAmqp,
-                FaultInjection.FaultCloseReason_Bye,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
         [TestMethod]
         public async Task Twin_ServiceSetsDesiredPropertyAndDeviceReceivesEvent_Mqtt()
         {
@@ -187,94 +95,6 @@ namespace Microsoft.Azure.Devices.E2ETests
         {
             await Twin_ServiceSetsDesiredPropertyAndDeviceReceivesEvent(Client.TransportType.Amqp_WebSocket_Only, SetTwinPropertyUpdateCallbackObsoleteHandlerAsync).ConfigureAwait(false);
         }
-
-        [Ignore] // TODO: #558
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Twin_DeviceDesiredPropertyUpdateTcpConnRecovery_Mqtt()
-        {
-            await Twin_DeviceDesiredPropertyUpdateRecovery(Client.TransportType.Mqtt_Tcp_Only, 
-                FaultInjection.FaultType_Tcp,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [Ignore] // TODO: #558
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Twin_DeviceDesiredPropertyUpdateTcpConnRecovery_MqttWs()
-        {
-            await Twin_DeviceDesiredPropertyUpdateRecovery(Client.TransportType.Mqtt_WebSocket_Only,
-                FaultInjection.FaultType_Tcp,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [Ignore] //TODO: #571
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Twin_DeviceDesiredPropertyUpdateTcpConnRecovery_Amqp()
-        {
-            await Twin_DeviceDesiredPropertyUpdateRecovery(Client.TransportType.Amqp_Tcp_Only, 
-                FaultInjection.FaultType_Tcp,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [Ignore] //TODO: #571
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Twin_DeviceDesiredPropertyUpdateTcpConnRecovery_AmqpWs()
-        {
-            await Twin_DeviceDesiredPropertyUpdateRecovery(Client.TransportType.Amqp_WebSocket_Only,
-                FaultInjection.FaultType_Tcp,
-                FaultInjection.FaultCloseReason_Boom,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [Ignore] // TODO: #558
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Twin_DeviceDesiredPropertyUpdateGracefulShutdownRecovery_Mqtt()
-        {
-            await Twin_DeviceDesiredPropertyUpdateRecovery(Client.TransportType.Mqtt_Tcp_Only,
-                FaultInjection.FaultType_GracefulShutdownMqtt,
-                FaultInjection.FaultCloseReason_Bye,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [Ignore] // TODO: #558
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Twin_DeviceDesiredPropertyUpdateGracefulShutdownRecovery_MqttWs()
-        {
-            await Twin_DeviceDesiredPropertyUpdateRecovery(Client.TransportType.Mqtt_WebSocket_Only,
-                FaultInjection.FaultType_GracefulShutdownMqtt,
-                FaultInjection.FaultCloseReason_Bye,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [Ignore] //TODO: #571
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Twin_DeviceDesiredPropertyUpdateGracefulShutdownRecovery_Amqp()
-        {
-            await Twin_DeviceDesiredPropertyUpdateRecovery(Client.TransportType.Amqp_Tcp_Only,
-                FaultInjection.FaultType_GracefulShutdownAmqp,
-                FaultInjection.FaultCloseReason_Bye,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        }
-
-        [Ignore] //TODO: #571
-        [TestMethod]
-        [TestCategory("IoTHub-FaultInjection")]
-        public async Task Twin_DeviceDesiredPropertyUpdateGracefulShutdownRecovery_AmqpWs()
-        {
-            await Twin_DeviceDesiredPropertyUpdateRecovery(Client.TransportType.Amqp_WebSocket_Only,
-                FaultInjection.FaultType_GracefulShutdownAmqp,
-                FaultInjection.FaultCloseReason_Bye,
-                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
-        } 
 
         [TestMethod]
         public async Task Twin_ServiceSetsDesiredPropertyAndDeviceReceivesItOnNextGet_Mqtt()
@@ -391,35 +211,6 @@ namespace Microsoft.Azure.Devices.E2ETests
             }
         }
 
-        private async Task Twin_DeviceReportedPropertiesRecovery(Client.TransportType transport, string faultType, string reason, int delayInSec)
-        {
-            var propName = Guid.NewGuid().ToString();
-            var props = new TwinCollection();
-
-            Func<DeviceClient, TestDevice, Task> testOperation = async (deviceClient, testDevice) =>
-            {
-                var propValue = Guid.NewGuid().ToString();
-                props[propName] = propValue;
-
-                await deviceClient.UpdateReportedPropertiesAsync(props).ConfigureAwait(false);
-
-                var deviceTwin = await deviceClient.GetTwinAsync().ConfigureAwait(false);
-                Assert.AreEqual<String>(deviceTwin.Properties.Reported[propName].ToString(), propValue);
-            };
-
-            await FaultInjection.TestErrorInjectionTemplate(
-                DevicePrefix,
-                TestDeviceType.Sasl,
-                transport,
-                faultType,
-                reason,
-                delayInSec,
-                FaultInjection.DefaultDurationInSec,
-                (d, t) => { return Task.FromResult<bool>(false); },
-                testOperation,
-                () => { return Task.FromResult<bool>(false); }).ConfigureAwait(false);
-        }
-
         private async Task<Task> SetTwinPropertyUpdateCallbackHandlerAsync(DeviceClient deviceClient, string expectedPropName, string expectedPropValue)
         {
             var propertyUpdateReceived = new TaskCompletionSource<bool>();
@@ -514,38 +305,6 @@ namespace Microsoft.Azure.Devices.E2ETests
 
                 await deviceClient.CloseAsync().ConfigureAwait(false);
             }
-        }
-  
-        private async Task Twin_DeviceDesiredPropertyUpdateRecovery(Client.TransportType transport, string faultType, string reason, int delayInSec)
-        {
-            RegistryManager registryManager = RegistryManager.CreateFromConnectionString(Configuration.IoTHub.ConnectionString);
-
-            var propName = Guid.NewGuid().ToString();
-            var props = new TwinCollection();
-
-            Func<DeviceClient, TestDevice, Task> testOperation = async (deviceClient, testDevice) =>
-            {
-                var propValue = Guid.NewGuid().ToString();
-                _log.WriteLine($"{nameof(Twin_DeviceDesiredPropertyUpdateRecovery)}: name={propName}, value={propValue}");
-
-                Task updateReceivedTask = await SetTwinPropertyUpdateCallbackHandlerAsync(deviceClient, propName, propValue).ConfigureAwait(false);
-
-                await Task.WhenAll(
-                    RegistryManagerUpdateDesiredPropertyAsync(testDevice.Id, propName, propValue),
-                    updateReceivedTask).ConfigureAwait(false);
-            };
-
-            await FaultInjection.TestErrorInjectionTemplate(
-                DevicePrefix,
-                TestDeviceType.Sasl,
-                transport,
-                faultType,
-                reason,
-                delayInSec,
-                FaultInjection.DefaultDurationInSec,
-                (d, t) => { return Task.FromResult<bool>(false); },
-                testOperation,
-                () => { return Task.FromResult<bool>(false); }).ConfigureAwait(false);
         }
 
         private async Task Twin_ServiceSetsDesiredPropertyAndDeviceReceivesItOnNextGet(Client.TransportType transport)
