+++ e2e/test/TwinFaultInjectionTests.cs
@@ -1,0 +1,320 @@
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using Microsoft.Azure.Devices.Client;
+using Microsoft.Azure.Devices.Shared;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using System;
+using System.Diagnostics.Tracing;
+using System.Threading.Tasks;
+
+namespace Microsoft.Azure.Devices.E2ETests
+{
+    [TestClass]
+    [TestCategory("IoTHub-E2E")]
+    public class TwinFaultInjectionTests : IDisposable
+    {
+        private readonly string DevicePrefix = $"E2E_{nameof(TwinFaultInjectionTests)}_";
+        private static TestLogging _log = TestLogging.GetInstance();
+
+        private readonly ConsoleEventListener _listener;
+
+        public TwinFaultInjectionTests()
+        {
+            _listener = TestConfig.StartEventListener();
+        }
+
+        [Ignore] // TODO: #558
+        [TestMethod]
+        [TestCategory("IoTHub-FaultInjection")]
+        public async Task Twin_DeviceReportedPropertiesTcpConnRecovery_Mqtt()
+        {
+            await Twin_DeviceReportedPropertiesRecovery(Client.TransportType.Mqtt_Tcp_Only,
+                FaultInjection.FaultType_Tcp,
+                FaultInjection.FaultCloseReason_Boom,
+                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
+        }
+
+        [Ignore] // TODO: #558
+        [TestMethod]
+        [TestCategory("IoTHub-FaultInjection")]
+        public async Task Twin_DeviceReportedPropertiesTcpConnRecovery_MqttWs()
+        {
+            await Twin_DeviceReportedPropertiesRecovery(Client.TransportType.Mqtt_WebSocket_Only,
+                FaultInjection.FaultType_Tcp,
+                FaultInjection.FaultCloseReason_Boom,
+                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
+        }
+
+        [TestMethod]
+        [TestCategory("IoTHub-FaultInjection")]
+        public async Task Twin_DeviceReportedPropertiesTcpConnRecovery_Amqp()
+        {
+            await Twin_DeviceReportedPropertiesRecovery(Client.TransportType.Amqp_Tcp_Only,
+                FaultInjection.FaultType_Tcp,
+                FaultInjection.FaultCloseReason_Boom,
+                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
+        }
+
+#if NETCOREAPP2_0
+        // TODO: #302 In NetCoreApp2.0 the test is failing with TimeoutException.
+        [Ignore]
+#endif
+        [TestMethod]
+        [TestCategory("IoTHub-FaultInjection")]
+        public async Task Twin_DeviceReportedPropertiesTcpConnRecovery_AmqpWs()
+        {
+            await Twin_DeviceReportedPropertiesRecovery(Client.TransportType.Amqp_WebSocket_Only,
+                FaultInjection.FaultType_Tcp,
+                FaultInjection.FaultCloseReason_Boom,
+                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
+        }
+
+        [Ignore] // TODO: #558
+        [TestMethod]
+        [TestCategory("IoTHub-FaultInjection")]
+        public async Task Twin_DeviceReportedPropertiesGracefulShutdownRecovery_Mqtt()
+        {
+            await Twin_DeviceReportedPropertiesRecovery(Client.TransportType.Mqtt_Tcp_Only,
+                FaultInjection.FaultType_GracefulShutdownMqtt,
+                FaultInjection.FaultCloseReason_Bye,
+                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
+        }
+
+        [Ignore] // TODO: #558
+        [TestMethod]
+        [TestCategory("IoTHub-FaultInjection")]
+        public async Task Twin_DeviceReportedPropertiesGracefulShutdownRecovery_MqttWs()
+        {
+            await Twin_DeviceReportedPropertiesRecovery(Client.TransportType.Mqtt_WebSocket_Only,
+                FaultInjection.FaultType_GracefulShutdownMqtt,
+                FaultInjection.FaultCloseReason_Bye,
+                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
+        }
+
+        [TestMethod]
+        [TestCategory("IoTHub-FaultInjection")]
+        public async Task Twin_DeviceReportedPropertiesGracefulShutdownRecovery_Amqp()
+        {
+            await Twin_DeviceReportedPropertiesRecovery(Client.TransportType.Amqp_Tcp_Only,
+                FaultInjection.FaultType_GracefulShutdownAmqp,
+                FaultInjection.FaultCloseReason_Bye,
+                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
+        }
+
+#if NETCOREAPP2_0
+        // TODO: #302 In NetCoreApp2.0 the test is failing with TimeoutException.
+        [Ignore]
+#endif
+        [TestMethod]
+        [TestCategory("IoTHub-FaultInjection")]
+        public async Task Twin_DeviceReportedPropertiesGracefulShutdownRecovery_AmqpWs()
+        {
+            await Twin_DeviceReportedPropertiesRecovery(Client.TransportType.Amqp_WebSocket_Only,
+                FaultInjection.FaultType_GracefulShutdownAmqp,
+                FaultInjection.FaultCloseReason_Bye,
+                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
+        }
+
+        [Ignore] // TODO: #558
+        [TestMethod]
+        [TestCategory("IoTHub-FaultInjection")]
+        public async Task Twin_DeviceDesiredPropertyUpdateTcpConnRecovery_Mqtt()
+        {
+            await Twin_DeviceDesiredPropertyUpdateRecovery(Client.TransportType.Mqtt_Tcp_Only,
+                FaultInjection.FaultType_Tcp,
+                FaultInjection.FaultCloseReason_Boom,
+                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
+        }
+
+        [Ignore] // TODO: #558
+        [TestMethod]
+        [TestCategory("IoTHub-FaultInjection")]
+        public async Task Twin_DeviceDesiredPropertyUpdateTcpConnRecovery_MqttWs()
+        {
+            await Twin_DeviceDesiredPropertyUpdateRecovery(Client.TransportType.Mqtt_WebSocket_Only,
+                FaultInjection.FaultType_Tcp,
+                FaultInjection.FaultCloseReason_Boom,
+                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
+        }
+
+        [Ignore] //TODO: #571
+        [TestMethod]
+        [TestCategory("IoTHub-FaultInjection")]
+        public async Task Twin_DeviceDesiredPropertyUpdateTcpConnRecovery_Amqp()
+        {
+            await Twin_DeviceDesiredPropertyUpdateRecovery(Client.TransportType.Amqp_Tcp_Only,
+                FaultInjection.FaultType_Tcp,
+                FaultInjection.FaultCloseReason_Boom,
+                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
+        }
+
+        [Ignore] //TODO: #571
+        [TestMethod]
+        [TestCategory("IoTHub-FaultInjection")]
+        public async Task Twin_DeviceDesiredPropertyUpdateTcpConnRecovery_AmqpWs()
+        {
+            await Twin_DeviceDesiredPropertyUpdateRecovery(Client.TransportType.Amqp_WebSocket_Only,
+                FaultInjection.FaultType_Tcp,
+                FaultInjection.FaultCloseReason_Boom,
+                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
+        }
+
+        [Ignore] // TODO: #558
+        [TestMethod]
+        [TestCategory("IoTHub-FaultInjection")]
+        public async Task Twin_DeviceDesiredPropertyUpdateGracefulShutdownRecovery_Mqtt()
+        {
+            await Twin_DeviceDesiredPropertyUpdateRecovery(Client.TransportType.Mqtt_Tcp_Only,
+                FaultInjection.FaultType_GracefulShutdownMqtt,
+                FaultInjection.FaultCloseReason_Bye,
+                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
+        }
+
+        [Ignore] // TODO: #558
+        [TestMethod]
+        [TestCategory("IoTHub-FaultInjection")]
+        public async Task Twin_DeviceDesiredPropertyUpdateGracefulShutdownRecovery_MqttWs()
+        {
+            await Twin_DeviceDesiredPropertyUpdateRecovery(Client.TransportType.Mqtt_WebSocket_Only,
+                FaultInjection.FaultType_GracefulShutdownMqtt,
+                FaultInjection.FaultCloseReason_Bye,
+                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
+        }
+
+        [Ignore] //TODO: #571
+        [TestMethod]
+        [TestCategory("IoTHub-FaultInjection")]
+        public async Task Twin_DeviceDesiredPropertyUpdateGracefulShutdownRecovery_Amqp()
+        {
+            await Twin_DeviceDesiredPropertyUpdateRecovery(Client.TransportType.Amqp_Tcp_Only,
+                FaultInjection.FaultType_GracefulShutdownAmqp,
+                FaultInjection.FaultCloseReason_Bye,
+                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
+        }
+
+        [Ignore] //TODO: #571
+        [TestMethod]
+        [TestCategory("IoTHub-FaultInjection")]
+        public async Task Twin_DeviceDesiredPropertyUpdateGracefulShutdownRecovery_AmqpWs()
+        {
+            await Twin_DeviceDesiredPropertyUpdateRecovery(Client.TransportType.Amqp_WebSocket_Only,
+                FaultInjection.FaultType_GracefulShutdownAmqp,
+                FaultInjection.FaultCloseReason_Bye,
+                FaultInjection.DefaultDelayInSec).ConfigureAwait(false);
+        }
+
+        private async Task Twin_DeviceReportedPropertiesRecovery(Client.TransportType transport, string faultType, string reason, int delayInSec)
+        {
+            var propName = Guid.NewGuid().ToString();
+            var props = new TwinCollection();
+
+            Func<DeviceClient, TestDevice, Task> testOperation = async (deviceClient, testDevice) =>
+            {
+                var propValue = Guid.NewGuid().ToString();
+                props[propName] = propValue;
+
+                await deviceClient.UpdateReportedPropertiesAsync(props).ConfigureAwait(false);
+
+                var deviceTwin = await deviceClient.GetTwinAsync().ConfigureAwait(false);
+                Assert.AreEqual<String>(deviceTwin.Properties.Reported[propName].ToString(), propValue);
+            };
+
+            await FaultInjection.TestErrorInjectionTemplate(
+                DevicePrefix,
+                TestDeviceType.Sasl,
+                transport,
+                faultType,
+                reason,
+                delayInSec,
+                FaultInjection.DefaultDurationInSec,
+                (d, t) => { return Task.FromResult<bool>(false); },
+                testOperation,
+                () => { return Task.FromResult<bool>(false); }).ConfigureAwait(false);
+        }
+
+        private async Task<Task> SetTwinPropertyUpdateCallbackHandlerAsync(DeviceClient deviceClient, string expectedPropName, string expectedPropValue)
+        {
+            var propertyUpdateReceived = new TaskCompletionSource<bool>();
+            string userContext = "myContext";
+
+            await deviceClient.SetDesiredPropertyUpdateCallbackAsync(
+                (patch, context) =>
+                {
+                    _log.WriteLine($"{nameof(SetTwinPropertyUpdateCallbackHandlerAsync)}: DesiredProperty: {patch}, {context}");
+
+                    try
+                    {
+                        Assert.AreEqual(expectedPropValue, patch[expectedPropName].ToString());
+                        Assert.AreEqual(userContext, context, "Context");
+                    }
+                    catch (Exception e)
+                    {
+                        propertyUpdateReceived.SetException(e);
+                    }
+                    finally
+                    {
+                        propertyUpdateReceived.SetResult(true);
+                    }
+
+                    return Task.FromResult<bool>(true);
+                }, userContext).ConfigureAwait(false);
+
+            return propertyUpdateReceived.Task;
+        }
+
+        private async Task RegistryManagerUpdateDesiredPropertyAsync(string deviceId, string propName, string propValue)
+        {
+            using (RegistryManager registryManager = RegistryManager.CreateFromConnectionString(Configuration.IoTHub.ConnectionString))
+            {
+                var twinPatch = new Twin();
+                twinPatch.Properties.Desired[propName] = propValue;
+
+                await registryManager.UpdateTwinAsync(deviceId, twinPatch, "*").ConfigureAwait(false);
+                await registryManager.CloseAsync().ConfigureAwait(false);
+            }
+        }
+
+        private async Task Twin_DeviceDesiredPropertyUpdateRecovery(Client.TransportType transport, string faultType, string reason, int delayInSec)
+        {
+            RegistryManager registryManager = RegistryManager.CreateFromConnectionString(Configuration.IoTHub.ConnectionString);
+
+            var propName = Guid.NewGuid().ToString();
+            var props = new TwinCollection();
+
+            Func<DeviceClient, TestDevice, Task> testOperation = async (deviceClient, testDevice) =>
+            {
+                var propValue = Guid.NewGuid().ToString();
+                _log.WriteLine($"{nameof(Twin_DeviceDesiredPropertyUpdateRecovery)}: name={propName}, value={propValue}");
+
+                Task updateReceivedTask = await SetTwinPropertyUpdateCallbackHandlerAsync(deviceClient, propName, propValue).ConfigureAwait(false);
+
+                await Task.WhenAll(
+                    RegistryManagerUpdateDesiredPropertyAsync(testDevice.Id, propName, propValue),
+                    updateReceivedTask).ConfigureAwait(false);
+            };
+
+            await FaultInjection.TestErrorInjectionTemplate(
+                DevicePrefix,
+                TestDeviceType.Sasl,
+                transport,
+                faultType,
+                reason,
+                delayInSec,
+                FaultInjection.DefaultDurationInSec,
+                (d, t) => { return Task.FromResult<bool>(false); },
+                testOperation,
+                () => { return Task.FromResult<bool>(false); }).ConfigureAwait(false);
+        }
+
+        public void Dispose()
+        {
+            Dispose(true);
+            GC.SuppressFinalize(this);
+        }
+
+        protected virtual void Dispose(bool disposing)
+        {
+        }
+    }
+}
